即时通讯基本功能初步实现 手机号正则匹配修正 退出登录功能实现

+++ app/src/main/java/com/example/mi_class/domain/Message.java
@@ -7,7 +7,7 @@ public class Message {
     public static final int TYPE_RECEIVE = 1;
     private String name;    //人名或课程名
     private String last_message;    //最后发送的信息
-    private int time;   //发送时间
+    private long time;   //发送时间
     private Bitmap head_portrait;   //头像
     private int type;   //类型
 
@@ -35,11 +35,11 @@ public class Message {
         this.last_message = last_message;
     }
 
-    public int getTime() {
+    public long getTime() {
         return time;
     }
 
-    public void setTime(int time) {
+    public void setTime(long time) {
         this.time = time;
     }
